Implementing "dreaming".
"Dreaming" is the "not immediately urgent" interrelation of information that finally makes a system smarter by letting it create further conclusions, and conclusions about "absurd" side-tracks. It happens by re-considering the past information (although the information is not pressing to be considered "now" - this "pressing" information is evaluated during "waking time"). - This gives a system a competitive advantage against "non-dreamers": it is able to generate further conclusions instead of "sitting idle", and it does not need to "urgently" make these conclusions during waking-time (i.e. it is OK to have fewer resources). [I actually believe these advantages apply to biological systems as well.]